Extends the test key to also select the first radio.
It is an ugly hack but it helps with debugging tremendously. The reason I have to sleep is to wait untill something is there to click, this race condition is impossible if you are the end user and you wait for the radio list to fill to click an item.